Effectful.Dispatch.Static.Primitive

Description

Primitive API for statically dispatched effects.

This module exposes internal implementation details of the Eff monad. Most of the time functions from Effectful.Dispatch.Static are sufficient.

Warning: playing the so called "type tetris" with functions from this module is not enough. Their misuse might lead to data races or internal consistency check failures, so make sure you understand what you're doing.

The environment

data Env (es :: [Effect])

A strict (WHNF), thread local, mutable, extensible record indexed by types of kind Effect.

Warning: the environment is a mutable data structure and cannot be simultaneously used from multiple threads under any circumstances.

In order to pass it to a different thread, you need to perform a deep copy with the cloneEnv funtion.

Offers very good performance characteristics for most often performed operations:

Extending: O(n), where n is the size of the effect stack.

Shrinking: O(1).

Indexing via (:>): O(1)

Modification of a specific element: O(1).

Getting a tail: O(1).

Cloning: O(N), where N is the size of the Storage.

Relinker

newtype Relinker :: (Effect -> Type) -> Effect -> Type where

A function for relinking Env objects stored in the handlers and/or making a deep copy of the representation of the effect when cloning the environment.

Constructors

Relinker :: ((forall es. Env es -> IO (Env es)) -> rep e -> IO (rep e)) -> Relinker rep e

dummyRelinker :: Relinker rep e

A dummy Relinker.

Representation of effects

type family EffectRep (d :: Dispatch) :: Effect -> Type

Internal representations of effects.

Extending and shrinking

consEnv

Arguments

:: EffectRep (DispatchOf e) e
The representation of the effect.

-> Relinker (EffectRep (DispatchOf e)) e

-> Env es

-> IO (Env (e : es))

Extend the environment with a new data type.

unconsEnv :: Env (e : es) -> IO ()

Shrink the environment by one data type.

Note: after calling this function e from the input environment is no longer usable.

Data retrieval and update

getEnv

Arguments

:: forall e es. e :> es

=> Env es
The environment.

-> IO (EffectRep (DispatchOf e) e)

Extract a specific data type from the environment.

putEnv

Arguments

:: forall e es. e :> es

=> Env es
The environment.

-> EffectRep (DispatchOf e) e

-> IO ()

Replace the data type in the environment with a new value (in place).

stateEnv

Arguments

:: forall e es a. e :> es

=> Env es
The environment.

-> (EffectRep (DispatchOf e) e -> IO (a, EffectRep (DispatchOf e) e))

-> IO a

Modify the data type in the environment and return a value (in place).

modifyEnv

Arguments

:: forall e es. e :> es

=> Env es
The environment.

-> (EffectRep (DispatchOf e) e -> IO (EffectRep (DispatchOf e) e))

-> IO ()

Modify the data type in the environment (in place).

Utils

emptyEnv :: IO (Env '[])

Create an empty environment.

cloneEnv :: Env es -> IO (Env es)

Clone the environment to use it in a different thread.

restoreEnv

Arguments

:: Env es
Destination.

-> Env es
Source.

-> IO ()

Restore the environment from its clone.

Since: 2.2.0.0

sizeEnv :: Env es -> IO Int

Get the current size of the environment.

tailEnv :: Env (e : es) -> IO (Env es)

Access the tail of the environment.