Effectful.Error.Dynamic

Description

The dynamically dispatched variant of the Error effect.

Note: unless you plan to change interpretations at runtime, it's recommended to use the statically dispatched variant, i.e. Effectful.Error.Static.

Effect

data Error e :: Effect where

Provide the ability to handle errors of type e.

Constructors

ThrowError :: e -> Error e m a

CatchError :: m a -> (CallStack -> e -> m a) -> Error e m a

Handlers

runError :: Eff (Error e : es) a -> Eff es (Either (CallStack, e) a)

Handle errors of type e (via Effectful.Error.Static).

runErrorWith

Arguments

:: (CallStack -> e -> Eff es a)
The error handler.

-> Eff (Error e : es) a

-> Eff es a

Handle errors of type e (via Effectful.Error.Static) with a specific error handler.

runErrorNoCallStack :: Eff (Error e : es) a -> Eff es (Either e a)

Handle errors of type e (via Effectful.Error.Static). In case of an error discard the CallStack.

runErrorNoCallStackWith

Arguments

:: (e -> Eff es a)
The error handler.

-> Eff (Error e : es) a

-> Eff es a

Handle errors of type e (via Effectful.Error.Static) with a specific error handler. In case of an error discard the CallStack.

Operations

throwError

Arguments

:: (HasCallStack, Error e :> es)

=> e
The error.

-> Eff es a

Throw an error of type e.

catchError

Arguments

:: (HasCallStack, Error e :> es)

=> Eff es a
The inner computation.

-> (CallStack -> e -> Eff es a)
A handler for errors in the inner computation.

-> Eff es a

Handle an error of type e.

handleError

Arguments

:: Error e :> es

=> (CallStack -> e -> Eff es a)
A handler for errors in the inner computation.

-> Eff es a
The inner computation.

-> Eff es a

The same as flip catchError, which is useful in situations where the code for the handler is shorter.

tryError

Arguments

:: (HasCallStack, Error e :> es)

=> Eff es a
The inner computation.

-> Eff es (Either (CallStack, e) a)

Similar to catchError, but returns an Either result which is a Right if no error was thrown and a Left otherwise.

Re-exports

type HasCallStack = ?callStack :: CallStack

Request a CallStack.

NOTE: The implicit parameter ?callStack :: CallStack is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.9.0.0

data CallStack

CallStacks are a lightweight method of obtaining a partial call-stack at any point in the program.

A function can request its call-site with the HasCallStack constraint. For example, we can define

putStrLnWithCallStack :: HasCallStack => String -> IO ()

as a variant of putStrLn that will get its call-site and print it, along with the string given as argument. We can access the call-stack inside putStrLnWithCallStack with callStack.

>>> :{
putStrLnWithCallStack :: HasCallStack => String -> IO ()
putStrLnWithCallStack msg = do
  putStrLn msg
  putStrLn (prettyCallStack callStack)
:}

Thus, if we call putStrLnWithCallStack we will get a formatted call-stack alongside our string.

>>> putStrLnWithCallStack "hello"
hello
CallStack (from HasCallStack):
  putStrLnWithCallStack, called at <interactive>:... in interactive:Ghci...

GHC solves HasCallStack constraints in three steps:

If there is a CallStack in scope -- i.e. the enclosing function has a HasCallStack constraint -- GHC will append the new call-site to the existing CallStack.

If there is no CallStack in scope -- e.g. in the GHCi session above -- and the enclosing definition does not have an explicit type signature, GHC will infer a HasCallStack constraint for the enclosing definition (subject to the monomorphism restriction).

If there is no CallStack in scope and the enclosing definition has an explicit type signature, GHC will solve the HasCallStack constraint for the singleton CallStack containing just the current call-site.

CallStacks do not interact with the RTS and do not require compilation with -prof. On the other hand, as they are built up explicitly via the HasCallStack constraints, they will generally not contain as much information as the simulated call-stacks maintained by the RTS.

A CallStack is a [(String, SrcLoc)]. The String is the name of function that was called, the SrcLoc is the call-site. The list is ordered with the most recently called function at the head.

NOTE: The intrepid user may notice that HasCallStack is just an alias for an implicit parameter ?callStack :: CallStack. This is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.8.1.0

getCallStack :: CallStack -> [([Char], SrcLoc)]

Extract a list of call-sites from the CallStack.

The list is ordered by most recent call.

Since: base-4.8.1.0

prettyCallStack :: CallStack -> String

Pretty print a CallStack.

Since: base-4.9.0.0