Effectful.Internal.Effect

Description

Type-safe indexing for Env.

This module is intended for internal use only, and may change without warning in subsequent releases.

Documentation

type Effect = (Type -> Type) -> Type -> Type

The kind of effects.

class (e :: Effect) :> (es :: [Effect]) where

A constraint that requires that a particular effect e is a member of the type-level list es. This is used to parameterize an Eff computation over an arbitrary list of effects, so long as e is somewhere in the list.

For example, a computation that only needs access to a mutable value of type Integer would have the following type:

State Integer :> es => Eff es ()

Minimal complete definition

Nothing

Methods

reifyIndex :: Int

Get the position of e in es.

Note: GHC is kind enough to cache these values as they're top level CAFs, so the lookup is amortized O(1) without any language level tricks.

type family xs :>> es :: Constraint where ...

Deprecated: Usage of (:>>) slows down GHC too much, so it will be removed in 3.0.0.0. See https://github.com/haskell-effectful/effectful/issues/52#issuecomment-1269155485 for more information.

Convenience operator for expressing that a function uses multiple effects in a more concise way than enumerating them all with (:>).

[E1, E2, ..., En] :>> es ≡ (E1 :> es, E2 :> es, ..., En :> es)

Equations

'[] :>> es = ()

(x : xs) :>> es = (x :> es, xs :>> es)

class KnownPrefix es => Subset (xs :: [Effect]) (es :: [Effect]) where

Provide evidence that xs is a subset of es.

Minimal complete definition

Nothing

Methods

subsetFullyKnown :: Bool

reifyIndices :: [Int]

class KnownPrefix (es :: [Effect]) where

Calculate length of a statically known prefix of es.

Methods

prefixLength :: Int

class IsUnknownSuffixOf (xs :: [Effect]) (es :: [Effect])

Require that xs is the unknown suffix of es.

Re-exports

type Type = TYPE LiftedRep

The kind of types with lifted values. For example Int :: Type.