Effectful.NonDet

Description

Provider of the Alternative and MonadPlus instance for Eff.

Effect

data NonDet :: Effect where

Provide the ability to use the Alternative and MonadPlus instance for Eff.

Since: 2.2.0.0

Constructors

Empty :: NonDet m a

(:<|>:) :: m a -> m a -> NonDet m a

data OnEmptyPolicy

Policy of dealing with modifications to thread local state in the environment in branches that end up calling the Empty operation.

Note: OnEmptyKeep is significantly faster as there is no need to back up the environment on each call to :<|>:.

Since: 2.2.0.0

Constructors

OnEmptyKeep
Keep modifications on Empty.

OnEmptyRollback
Rollback modifications on Empty.

Handlers

runNonDet :: OnEmptyPolicy -> Eff (NonDet : es) a -> Eff es (Either CallStack a)

Run the NonDet effect with a given OnEmptyPolicy.

Note: :<|>: executes the second computation if (and only if) the first computation calls Empty.

Since: 2.2.0.0

Utils

emptyEff :: (HasCallStack, NonDet :> es) => Eff es a

Specialized version of empty with the HasCallStack constraint for tracking purposes.

Since: 2.2.0.0

sumEff :: (HasCallStack, Foldable t, NonDet :> es) => t (Eff es a) -> Eff es a

Specialized version of asum with the HasCallStack constraint for tracking purposes.

Since: 2.2.0.0

Re-exports

class Applicative f => Alternative (f :: Type -> Type) where

A monoid on applicative functors.

If defined, some and many should be the least solutions of the equations:

some v = (:) <$> v <*> many v

many v = some v <|> pure []

Minimal complete definition

empty, (<|>)

Methods

empty :: f a

The identity of <|>

(<|>) :: f a -> f a -> f a infixl 3

An associative binary operation

some :: f a -> f [a]

One or more.

many :: f a -> f [a]

Zero or more.

type HasCallStack = ?callStack :: CallStack

Request a CallStack.

NOTE: The implicit parameter ?callStack :: CallStack is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.9.0.0

data CallStack

CallStacks are a lightweight method of obtaining a partial call-stack at any point in the program.

A function can request its call-site with the HasCallStack constraint. For example, we can define

putStrLnWithCallStack :: HasCallStack => String -> IO ()

as a variant of putStrLn that will get its call-site and print it, along with the string given as argument. We can access the call-stack inside putStrLnWithCallStack with callStack.

>>> :{
putStrLnWithCallStack :: HasCallStack => String -> IO ()
putStrLnWithCallStack msg = do
  putStrLn msg
  putStrLn (prettyCallStack callStack)
:}

Thus, if we call putStrLnWithCallStack we will get a formatted call-stack alongside our string.

>>> putStrLnWithCallStack "hello"
hello
CallStack (from HasCallStack):
  putStrLnWithCallStack, called at <interactive>:... in interactive:Ghci...

GHC solves HasCallStack constraints in three steps:

If there is a CallStack in scope -- i.e. the enclosing function has a HasCallStack constraint -- GHC will append the new call-site to the existing CallStack.

If there is no CallStack in scope -- e.g. in the GHCi session above -- and the enclosing definition does not have an explicit type signature, GHC will infer a HasCallStack constraint for the enclosing definition (subject to the monomorphism restriction).

If there is no CallStack in scope and the enclosing definition has an explicit type signature, GHC will solve the HasCallStack constraint for the singleton CallStack containing just the current call-site.

CallStacks do not interact with the RTS and do not require compilation with -prof. On the other hand, as they are built up explicitly via the HasCallStack constraints, they will generally not contain as much information as the simulated call-stacks maintained by the RTS.

A CallStack is a [(String, SrcLoc)]. The String is the name of function that was called, the SrcLoc is the call-site. The list is ordered with the most recently called function at the head.

NOTE: The intrepid user may notice that HasCallStack is just an alias for an implicit parameter ?callStack :: CallStack. This is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.8.1.0

getCallStack :: CallStack -> [([Char], SrcLoc)]

Extract a list of call-sites from the CallStack.

The list is ordered by most recent call.

Since: base-4.8.1.0

prettyCallStack :: CallStack -> String

Pretty print a CallStack.

Since: base-4.9.0.0