Effectful.Reader.Dynamic

Description

The dynamically dispatched variant of the Reader effect.

Note: unless you plan to change interpretations at runtime, it's recommended to use the statically dispatched variant, i.e. Effectful.Reader.Static.

Effect

data Reader r :: Effect where

Constructors

Ask :: Reader r m r

Local :: (r -> r) -> m a -> Reader r m a

Handlers

runReader

Arguments

:: r
The initial environment.

-> Eff (Reader r : es) a

-> Eff es a

Run the Reader effect with the given initial environment (via Effectful.Reader.Static).

withReader

Arguments

:: (r1 -> r2)
The function to modify the environment.

-> Eff (Reader r2 : es) a
Computation to run in the modified environment.

-> Eff (Reader r1 : es) a

Execute a computation in a modified environment.

Since: 1.1.0.0

Operations

ask :: (HasCallStack, Reader r :> es) => Eff es r

Fetch the value of the environment.

asks

Arguments

:: (HasCallStack, Reader r :> es)

=> (r -> a)
The function to apply to the environment.

-> Eff es a

Retrieve a function of the current environment.

asks f ≡ f <$> ask

local

Arguments

:: (HasCallStack, Reader r :> es)

=> (r -> r)
The function to modify the environment.

-> Eff es a

-> Eff es a

Execute a computation in a modified environment.

runReader r (local f m) ≡ runReader (f r) m