Effectful.Reader.Static

Description

Support for access to a read only value of a particular type.

Effect

data Reader r :: Effect

Provide access to a strict (WHNF), thread local, read only value of type r.

Handlers

runReader

Arguments

:: r
The initial environment.

-> Eff (Reader r : es) a

-> Eff es a

Run a Reader effect with the given initial environment.

withReader

Arguments

:: (r1 -> r2)
The function to modify the environment.

-> Eff (Reader r2 : es) a
Computation to run in the modified environment.

-> Eff (Reader r1 : es) a

Execute a computation in a modified environment.

Since: 1.1.0.0

Operations

ask :: Reader r :> es => Eff es r

Fetch the value of the environment.

asks

Arguments

:: Reader r :> es

=> (r -> a)
The function to apply to the environment.

-> Eff es a

Retrieve a function of the current environment.

asks f ≡ f <$> ask

local

Arguments

:: Reader r :> es

=> (r -> r)
The function to modify the environment.

-> Eff es a

-> Eff es a

Execute a computation in a modified environment.

runReader r (local f m) ≡ runReader (f r) m