Effectful.Internal.Utils

Documentation

inlineBracket :: IO a -> (a -> IO b) -> (a -> IO c) -> IO c

Version of bracket with an INLINE pragma to work around https://gitlab.haskell.org/ghc/ghc/-/issues/22824.

Utils for ThreadId

weakThreadId :: ThreadId -> Int

Get an id of a thread that doesn't prevent its garbage collection.

eqThreadId :: ThreadId -> ThreadId -> Bool

Utils for Any

type family Any :: k where ...

The type constructor Any is type to which you can unsafely coerce any lifted type, and back. More concretely, for a lifted type t and value x :: t, unsafeCoerce (unsafeCoerce x :: Any) :: t is equivalent to x.

toAny :: a -> Any

fromAny :: Any -> a

Strict IORef

data IORef' a

A strict variant of IORef.

newIORef' :: a -> IO (IORef' a)

readIORef' :: IORef' a -> IO a

writeIORef' :: IORef' a -> a -> IO ()

Strict MVar

data MVar' a

A strict variant of MVar.

toMVar' :: MVar a -> IO (MVar' a)

newMVar' :: a -> IO (MVar' a)

readMVar' :: MVar' a -> IO a

modifyMVar' :: MVar' a -> (a -> IO (a, r)) -> IO r

modifyMVar_' :: MVar' a -> (a -> IO a) -> IO ()

Unique

data Unique

A unique with no possibility for CAS contention.

Credits for this go to Edward Kmett.

newUnique :: IO Unique

CallStack

thawCallStack :: CallStack -> CallStack